# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "",
        "version": "... BOOKPERFECT API DESIGNED BY HOTELISTAN COMPANY ..."
    },
    "servers": [
        {
            "url": "https://cms.hitrooms.com/rest-api"
        }
    ],
    "paths": {
...
        "/get__transport_supplier_option_Code/{supplierid}/{transportid}/{optioncode}": {
            "get": {
                "tags": [
                    "@Contracts"
                ],
                "summary": "GET TRANSPORT BY SUPPLIERID AND TRANSPORTID AND OPTIONCODE",
                "operationId": "12",
                "parameters": [
                    {
                        "name": "supplierid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "transportid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "optioncode",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/get__accommodations__by__id/{accommodationId}": {
            "get": {
                "tags": [
                    "@Web content"
                ],
                "summary": "GET ACCOMMODATION BY ACCOMMODATIONID",
                "operationId": "22",
                "parameters": [
                    {
                        "name": "accommodationId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
        "/get__destination/{city}": {
            "get": {
                "tags": [
                    "@Web content"
                ],
                "summary": "GET DESTINATION BY CITY",
                "operationId": "28",
                "parameters": [
                    {
                        "name": "city",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/get__destination__by_code/{code}": {
            "get": {
                "tags": [
                    "@Web content"
                ],
                "summary": "GET ALL DESTINATION",
                "operationId": "31",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/get__Booking_all/{from}/{to}": {
            "get": {
                "tags": [
                    "@Booking"
                ],
                "summary": "GET ALL BOOKING",
                "operationId": "200",
                "parameters": [
                    {
                        "name": "from",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "to",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lang",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
        "/get__hotel_supplierid_provider_code/{supplierid}/{providercode}": {
            "get": {
                "tags": [
                    "@Contracts"
                ],
                "summary": "GET HOTEL BY SUPPLIERID AND PROVIDERCODE",
                "operationId": "3",
                "parameters": [
                    {
                        "name": "supplierid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "providercode",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
        "/get__TripIdeasCategoriesHotelten__by__name/{city}": {
            "get": {
                "tags": [
                    "@TripIdeasCategories"
                ],
                "summary": "GET 10 TOP HOTEL ID",
                "operationId": "61",
                "parameters": [
                    {
                        "name": "city",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
        "/supplierid/{supplierid}": {
            "get": {
                "tags": [
                    "@Contracts"
                ],
                "summary": "GET SUPPLIER BY SUPPLIERID",
                "operationId": "5",
                "parameters": [
                    {
                        "name": "supplierid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
        "/get__ticket_supplierid/{supplierid}": {
            "get": {
                "tags": [
                    "@Contracts"
                ],
                "summary": "GET TICKET BY SUPPLIERID",
                "operationId": "6",
                "parameters": [
                    {
                        "name": "supplierid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    }
}